Remove MBRange::iterator and MBRange::reverse_iterator.  Replace them with:  typedef const_iterator iterator;        typedef const_reverse_iterator reverse_iterator;
Removed non-const versions of begin, end, rbegin, rend, find, and
lower_bound.

Justification for above changes:

a) Setting a value through the iterator should not be allowed because
   it would allow changing a particular handle in a sequence such
   that the sequence is no longer ordered, violating the design of
   MBRange.  See std::set for a similar case in the STL.

b) The non-const iterators didn't allow changing of the range data,
   but rather allowed changing of a local variable in the iterator.
   This 1) violates the design of iterators in general, 2) isn't
   particularly useful, and 3) could potentially leave the iterator
   in an invalid state.

c) Removes a lot of duplicate code from MBRange.



Changed the implementation of const_reverse_iterator to be a simple
wrapper around a const_iterator which does the inverse of the stepping
operators on the wrapped iterator.

Justification for above change:

a) Data structure and iterator implementation support doing so.

b) Combined with first block of changes, reduces the number of
   actual iterator implementations from four to one.



Removed copy constructors from iterators.

Justification for change:

Compiler will provide default ones that do the same thing.  It was
unnecessary code and some compilers may potentially inline the
compiler-generated ones better.  I.e. there is a very small
possibility of a performance increase with no chance of a loss of
performance.



Added += and -= operators to iterators.

Justification for change:

I needed an method to retrieve the Nth entry from a range where N
is potentially a very large value.  Calling operator++ in a FOR
loop would be too inefficient.



Changed MBEntityHandle value assigned to dummy list node from
~0 to 0.

Justification for change:

Zero is always an invalid handle value.  ~0 may not be on 32-bit
platforms and certainly won't be on 64-bit ones.



Overall, above changes add a bit of new functionality and still
reduce the combined size of MBRange.hpp and MBRange.cpp by more
than 300 lines.

Passed all MOAB tests.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@464 6162379e-bd28-0410-9a7d-b7f4dcfcad3c