Add run-time registration of file readers/writers (not runtime loading of plugins yet, just the ability for an app to register its own reader and/or writer.)
Also add selection of reader/writer based on file extension.  If no writer
exists for the file  extension, an exodus file will be written. The reader
for the file extension will be used if there is one.  Otherwise all readers
are tried.  If there is a reader for a given file extension and that reader
fails, other readers are NOT tried.

For file-extension matching, a case-sensitive search is tried first.
If no match is found, a case-insensitive search is tried.

Add MBReaderIface and MBWriterIface as reader/writer interface definition
and change all readers/writers to implement above interfaces.

Change exodus writer to export all mesh as one block if no blocks are
defined.  Old behavior was to fail in this case.

Quick hack to .cub file reader to make it not a singleton.  Tested with
simple mesh-a-brick test case.

All built-in readers and writers are registered in the MBReaderWriterSet
constructor.  File extensions and such for built-in readers are hard-coded
there.

moab_test succeeds for the above changes.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@473 6162379e-bd28-0410-9a7d-b7f4dcfcad3c