Fixing FOC operation to work in cases where you're splitting/joining two blind chords.
Attempting to get code to work for exporting mesh to vtk after dual is generated.

Adding -rpath option for building converter.

Fixing code to draw sheets to treat blind sheets with > 2 chords.

Adding -d option to qvdual, to build dual automatically after
importing.  Also changing the way sheets are drawn after operations,
to offset them in the x direction and draw the most relevant sheet
last so it's on top.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@656 6162379e-bd28-0410-9a7d-b7f4dcfcad3c