Fix bug in building of tests, introduced with my previous checkin.
When compiling with mpiCC/mpicc, attempt to get name of real compiler
for determining special compiler flags.

Fix recursive build of tests in tests/* subdirs.

Fix makefile target to unzip test/mb_big_test.g.gz automatically when
running moab_test such that it work in Sun's make.

Always compile source in mhdf/src with -DNDEBUG.  Otherwise some checks
for open HDF5 handes are done.  This is a) inefficient and b) fails
with different versions of the HDF5 library due to bugs in the library.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@691 6162379e-bd28-0410-9a7d-b7f4dcfcad3c