Add missing #include
Change errors to warnings in qvdual and mbchaco configure scripts, because
script must succeed for "make dist" and other targets regardless of whether
or not the build of the tool will succeed.

Always do recursive configure on qvdual and mbchaco, even if tool is not
to be built.  This must be done to ensure that Makefiles are created for
targets such as "make dist".

Change syntax for recursive automake stuff in tools/Makefile.am to make
sure that "make dist" works for disabled tools.

Check in some old fixes to tools/qvdual/Makfile.am that never got checked in.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@721 6162379e-bd28-0410-9a7d-b7f4dcfcad3c