Fix bug in HDF5 writer: segfault writing tags containing arrays of (more than one) entity handle.
Fix bug in HDF5 writer: invalid data written when a subset of the
mesh is to be written and one or more entities in that subset have
adjacencies to entities not in that subset.

Remove inefficient code for initializing a tag value for all entities.
Just specify the value as the default when creating the tag.

Add #define to enable writing node->entity adjacencies.  The previous code
never wrote node->entity adjacencies (I don't recall why.)  Neither does
the code after this change, but writing of those adjacencies can now be
enabled with a #define.

Use simple sorted vector instead of RangeTree for O(n ln n) searches
when writing HDF5 files, and remove RangeTree class.  This requires more
memory, but it is a lot faster.  This change was made when I was last
working on the parallel writer, but never got checked in.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@728 6162379e-bd28-0410-9a7d-b7f4dcfcad3c