Fix incorrect implementation of increment operators for MBEntityType.
It is not safe to assume that the size of an enum type is the same as an
integer.  Even if sizeof(..) returns the same value for both, the compiler,
when optimizing, need not make temporary values any larger than necessary
to contain the range of the enum.

The old implementation resulted in incorrect behavior when compiling
with g++ 4.1 on a 32-bit little-endian platform.

Remove previously checked-in workaround for this bug in WriteHDF5.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@789 6162379e-bd28-0410-9a7d-b7f4dcfcad3c