Change the storage of handle->meshset mapping and new set ID allocation. This change restores the old behavior, where the ID used in new entity set handles is always increasing (don't re-use handles from deleted sets.) This gives better performance for dense tags on entity sets.  This change might make deleting entity sets a little slower.  Creating entity sets will probably be faster than either the original code or my previous implementation.  Accessing entity sets should take approx. the same amount of time as my previous implementation.
MeshSetManager now maintains a 3-D array for finding MeshSet*'s given
entity handles.  For 32-bit handles, the handle is divided as follows:

   28  24      16       8   4   0
 ...|...|...|...|...|...|...|...|
 <MS><--i---><---j----><---k---->

i, j, k are indices into the 3-D array.  j and k are each 10-bits,
which result in arrays exactly 1 page (4k) in size on 32-bit machines
(2 4k pages on 64-bit machines.)  i is 8, so the size of the array
is 256x1024x1024.  The 256-entry first level of the array is embedded
in the MeshSetManager class.  The 1024-entry secondary and tertiary
level arrays are allocated as needed and freed when they become empty.
The performance loss during deletion is due to the choice to not store
a count of the number of occupied slots in the secondary and tertiary
arrays, necessitating as search of the array when mesh sets are deleted to
ascertain if it is empty and should be deleted.  Allocation of secondary
and tertiary arrays is done using malloc/free rather than new/delete to
avoid the C++ array overhead, which would mess up the nice page alignment.

For 64-bit handles, the handle is split as follows

   60  56      48      40      32      24      16       8   4   0
 ...|...|...|...|...|...|...|...|...|...|...|...|...|...|...|...|
 <MS><--------i---------><--------j---------><--------k--------->

Each index is 20 bits, so the matrix is 1Mx1Mx1M.  The allocation of 4MB
arrays might be problematic, and it is probably wasteful at the primary
and secondary levels to have 1M-entry arrays.  However, it is a fairly
simple change from the 32-bit handle case, and the overall wasted space
is at most 8MB, which hopefully isn't too much on a 64-bit machine.
A more efficient 64-bit implementation would be to use 4 (or perhaps 5)
levels (4-D matrix).  It would be even more efficient and practical to
just impose a limit on the number of sets smaller than the available
ID space.  It is unlikely that any application will be creating anywhere
near 10^18 entity sets.  However, the hole in the ID space would need
to be introduced in the middle.  The upper range is used to store the
parallel processor ID, so that portion must be able to vary.  Placing the
hole at the bottom would break dense tags on sets.

 MeshSetManager.cpp |  243 ++++++++++++++++++++++++++++++++---------------------
 MeshSetManager.hpp |   31 ++++--
 2 files changed, 171 insertions(+), 103 deletions(-)


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@848 6162379e-bd28-0410-9a7d-b7f4dcfcad3c