Rename config.make to moab.make so that, when installed, it doesn't conflict with other packages' files (e.g. CGM.)
Install moab.make in $libdir rather than $includedir because only
platform-independent stuff is supposed be installed in $includedir
(should be able to install builds for different platforms and have them
all share one $includedir).

Remove code from configure.in to put recursively expanded values of
$includedir and $libdir in config.make.  Instead, use an install hook
to append the correct values to the file after it is installed.  This
ensures that the values are right if an alternate installation directory
was specified to make after configure is run (e.g. 'make prefix=/foo').
It also eliminates the need for special shell code to recursively expand
the variables, as make handles that for us.  This is a little more messy:
some of the code for generating moab.make is in Makefile.am rather than
all of it being in moab.make.in.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@883 6162379e-bd28-0410-9a7d-b7f4dcfcad3c