Fix bug writing default and/or global/mesh tags containing more than one MBEntityHandle.  Need to convert all values to file IDs, not just the first one.
WriteHDF5 skips writing tags containing entity handles that are either
invalid or that point to entities that are not also being written to the
file.  Allow special case of NULL MBEntityHandle (zero value.)  For tags
containing multiple entity handles, write the tag if *any* of the values
are valid, filling in any invalid ones with the NULL handle.

Add common function for converting tag data containing entity handles to
file ids and call from all the places in WriteHDF5 where that was being done.

Fix reading of default and/or mesh values for tags that are arrays of
MBEntityHandles.  Need to convert all values from file id to entity handle,
not just the first one.

When reading file, treat a zero file ID as a special case (NULL MBEntityHandle).

Fix horrendously bad coding decision by me (Jason): do not have two long-ish
variable names that differ by only the *case* of *one* of the letters (and
not even the first one.)

 ReadHDF5.cpp  |   47 +++++++++++++++++++---------------
 WriteHDF5.cpp |   80 +++++++++++++++++++++++++++++++++++++---------------------
 2 files changed, 79 insertions(+), 48 deletions(-)


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@892 6162379e-bd28-0410-9a7d-b7f4dcfcad3c