Make it easier to debug HDF5 writer with valgrind.  Because the HDF5 library caches data to be written later (often on close) it is difficult to determine where uninitialized data originated from.  Further, because WriteHDF5 re-uses the same buffer for most large IO, valgrind does not detect writes of uninitialized data because it sees the data as being initialized to whatever values were previously written.
When building with --enable-debug and the valgrind/memcheck.h header is
present, wrap calls to H5Dwrite with calls to notify valgrind that data
must be initialized before entering HDF5write and that the data should
be treated as uninitialized after leaving the HDF5write call.  Valgrind
calls have no effect if code is not running inside of valgrind.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@2014 6162379e-bd28-0410-9a7d-b7f4dcfcad3c