Generate targets for installed makefile automatically using contents of check_PROGRAMS from Makefile.am.
Note:  Assumes that a) all examples are listed in check_PROGRAMS and
       b) that the source file name is the executable name with a
       .cpp suffix.


The resulting file is identical to the previous one with manually specified
targets.  It is just generated automatically now, removing the need to
specify example targets in both Makefile.am and examples.make.in, as long
as the conventions above are followed.




git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3974 6162379e-bd28-0410-9a7d-b7f4dcfcad3c