Update DagMC's update_coords tool to compile under the new -pedantic regime.
Convert old variable-length arrays to vectors.  In the function skin_hex_elems,
there is a multidimensional variable-length array for which this conversion
is not trivial (because the array participates in qsort calls).  But it seems
that skin_hex_elems and its qsort comparator functions are not actually called
by anything, or exported in headers, so I have just wrapped them in an #if 0.
I'll plan to remove this dead code once I talk to its author.

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4295 6162379e-bd28-0410-9a7d-b7f4dcfcad3c