Bugfix to implicit complement handling
Simplify build_indices() by including the implicit complement's
handle in the 'vols' parameter before calling the function.  This is
in keeping with r4879 and makes the implicit complement even more like a
"normal" volume.

This fixes a rare problem introduced by r4879, occuring when
a file with an existing implicit complement was loaded: the
complement volume would appear in the indices twice.  That
situation led to errors in dag-mcnp5. Thanks to Patrick Snouffer
for reporting the bug.

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5047 6162379e-bd28-0410-9a7d-b7f4dcfcad3c