Add rudimentary Makefile support for latex documents.
This is fairly basic in that it does not include configure support for
detecting the necessary tools (latex, dia, dvipdf, etc.) or the logic
for handling the case where said tools are not available. So this change
is for developer convenience only.  Continue to check in generated PDFs
so that automake doesn't try to re-build them for users.

config.tex(.in) (copied from Mesquite source) contains the magic for 
doing latex in a separate build directory.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5121 6162379e-bd28-0410-9a7d-b7f4dcfcad3c