Changes in Skinner API
A file set is added in the list of arguments for the skinner.
This meshset usually is the file set used to load the mesh, and it
should also contain the source entities.
It was added to avoid confusion when multiple files are loaded in the same instance, for example
in coupler scenarios