create coarse and fine mesh from homme fvm run
the input is a file, but it can come directly
from homme run.
Still need to pass the allocatable pointers (in Fortran)
towards c++ code.

this is just the arrival mesh, the departure mesh will be
conputed later, at each time step

the fine mesh is not yet merged
there are some issues with the corner ids, they are
not matching well.
Or is there a bug in my code?
that is impossible :)