change the file_id_tag type again
So during reading a hdf5 file in parallel, we create a no-name tag that will be used to identify
uniquely the vertices, using the hdf5 id from the file. This hdf5 file id is usually a long,
 but we do not take any chances, we create an opaque tag of size (long), which is usually 8 on 64 bit
machines and 4 on 32 bit machines.
Before, we used an integer tag, and this limited the correct reading of a file with more than
1B vertices. Then we used a handle tag, but what we really need is a size(long) tag. Crystal
router needs a long for that unique identifier.

Now, we convert the file id to "long" always
So this will work even when we use 32 bit handles on 64 bit machines.