Several updates to the interface to make it consistent.
Immutable input params are passed by value. Use pre-processor
defines to change type names so that we can control some of
these at configure time.

- Change the return code to plain int. By changing return code,
the driver does not depend on moab, it compiles just with imoab.h
include.
- Also include some ifdef guards
- imoab.cpp will contain the implementation

Note: it depends on mhdf header, which is not part of installation
it is copied here temporarily

Former-commit-id: f8fe4ab67f1991b62428c7dcb1856d9ce49fe741