Implement more API methods related to parallel queries. A simple version of the driver has also been added.
- more global variables, that store file ids, pcomm, mapping between pid and STRING
entities are ghosted, but the material sets/ neumann sets are not filled yet
with ghost entities

- ghost addl entities too. Will pass 3 for addl entities when ghosting
it means that edges and faces will be ghosted too. It still does not
resolve the issue of populating the classic sets. So the Neumann sets
and material sets, after ghosting, will not contain any ghost entities.
Will need an extra step / call to solve this problem.
- Add support for deregistering the application

should delete the relevant entities
it is failing now to delete entities from the file set
1) addl_ents in ghosting are not added to the file set
2) not sure if skin entities are added either


Former-commit-id: 136c54ea19955dbf517a33c466f6de10ef1fdea9