additional sub-entities when ghosting
when we use file set for resolve shared entities and
ghosting, we add the new entities to the file set
(for example skin entities, faces and edges), and all ghost
entities
when we use addl_ents flag during ghosting, we first create
the subentities on current task, based on number of layers.
those entities created before actual ghosting takes place,
need to be added to the file set too.