test_entsetFB leaks
free some more memory
entsetFB is almost the same as entset , except the
geometry engine is FBEngine from moab, instead of cgma
test_entsetFB is almost free, compared to test_entset, in which
a large chunk of memory is still being pointed to at.
It seems to be CGM, in which the CGMApp::shutdown() deletes
all singletons.