augment fix: remove conservative estimate
When we initialize the tuple list for routing,
we estimate the capacity based on the number of
shared entities.
Before, we were using 30*sharedEnts.size(), which is probably
too high.

instead, start with a lower estimate for the
TupleList (sharedEnts.size())  and resize if
we are in danger of exceeding the capacity